+++ visor/visor.go
@@ -107,7 +107,7 @@ func (v *Visor) Run(ctx context.Context) error {
 
 		v.conf.SetHasDataNode(runConf.DataNode != nil)
 
-		client := v.clientFactory.GetClient(
+		c := v.clientFactory.GetClient(
 			runConf.Vega.RCP.SocketPath,
 			runConf.Vega.RCP.HTTPPath,
 		)
@@ -118,7 +118,12 @@ func (v *Visor) Run(ctx context.Context) error {
 		maxNumRestarts := v.conf.MaxNumberOfRestarts()
 		restartsDelay := time.Second * time.Duration(v.conf.RestartsDelaySeconds())
 
-		v.log.Info("Starting binaries")
+		if isRestarting {
+			v.log.Info("Restarting binaries")
+		} else {
+			v.log.Info("Starting binaries")
+		}
+
 		binRunner := NewBinariesRunner(
 			v.log,
 			v.conf.CurrentFolder(),
@@ -151,14 +156,14 @@ func (v *Visor) Run(ctx context.Context) error {
 
 				break CheckLoop
 			case <-upgradeTicker.C:
-				upStatus, err := client.UpgradeStatus(ctx)
+				upStatus, err := c.UpgradeStatus(ctx)
 				if err != nil {
 					// Binary has not started yet - waiting for first startup
 					if numOfRestarts == 0 {
 						if numOfUpgradeStatusErrs > maxNumberOfFirstConnectionRetries {
 							return failedToGetStatusErr(maxNumberOfFirstConnectionRetries, err)
 						}
-					} else { // Binary has been started already. Somethig has failed after the startup
+					} else { // Binary has been started already. Something has failed after the startup
 						if numOfUpgradeStatusErrs > maxUpgradeStatusErrs {
 							return failedToGetStatusErr(maxUpgradeStatusErrs, err)
 						}